requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
distro==1.9.0
Django==6.0.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
Flask==3.1.2
fonttools==4.61.0
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.45
groq==1.1.1
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mysql-connector-python==9.5.0
narwhals==2.13.0
numpy==2.3.3
openai==2.26.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
psycopg2-binary==2.9.11
pyarrow==21.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydyf==0.12.1
pyparsing==3.2.5
pypdf==6.8.0
pyphen==0.17.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
referencing==0.37.0
reportlab==4.4.5
requests==2.32.5
rpds-py==0.30.0
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.5
streamlit==1.51.0
tenacity==9.1.2
tinycss2==1.5.1
tinyhtml5==2.0.0
tkcalendar==1.6.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
weasyprint==68.0
webencodings==0.5.1
Werkzeug==3.1.4
whitenoise==6.12.0
xlsxwriter==3.2.9
zopfli==0.4.0
